refactor: implement `useMemo` when filtering data
