+++ src/components/organisms/tableJadwalSholat/index.tsx
@@ -1,6 +1,6 @@
 import { Waktu } from "@/interfaces";
 import { tableJadwalSholatList } from "@/utils/data";
-import { matchDate } from "@/utils/date";
+import { matchDate } from "@/helpers/formatDate";
 
 const TableJadwalSholat = ({ waktu }: Waktu) => {
   return (
